+++ src/server/routes/services/{serviceId}/resources/{environment}/page.njk
