+++ src/server/routes/services/{serviceId}/resources/{environment}/route.js
@@ -1,3 +1,5 @@
+import { formatText } from '#config/nunjucks/filters/filters.js'
+import { fetchResources } from '#server/services/helpers/fetch/fetch-resources.js'
 import { serviceParamsValidation } from '#server/services/helpers/schema/service-params-validation.js'
 import Boom from '@hapi/boom'
 
@@ -11,4 +13,51 @@ export const options = {
   }
 }
 
-export default async function (request) {}
+export default async function (request) {
+  const { environment } = request.params
+  const { entity } = request.app
+  const serviceName = entity.name
+
+  const team = entity?.teams?.at(0)
+  const teamId = team?.teamId
+  const formattedEnvironment = formatText(environment)
+
+  const resources = await fetchResources(entity.name, environment)
+
+  if (!resources) throw new Error('Failed to load resources')
+
+  const hasNoResources = !Object.entries(resources).some(
+    ([_, items]) => items?.length
+  )
+
+  const debugView = request.query.debug ?? false
+  const template = debugView
+    ? 'services/service/resources/views/debug/environment'
+    : 'services/service/resources/views/environment'
+
+  return {
+    pageTitle: `${serviceName} - Resources - ${formattedEnvironment}`,
+    entity,
+    teamId,
+    environment,
+    resources,
+    hasNoResources,
+    breadcrumbs: [
+      {
+        text: 'Services',
+        href: '/services'
+      },
+      {
+        text: serviceName,
+        href: `/services/${serviceName}`
+      },
+      {
+        text: 'Resources',
+        href: `/services/${serviceName}/resources`
+      },
+      {
+        text: formattedEnvironment
+      }
+    ]
+  }
+}
