+++ src/server/services/service/routes.js
@@ -5,7 +5,6 @@ import { serviceTerminal } from './terminal/routes.js'
 import { serviceAutomations } from './automations/routes.js'
 import { serviceNotifications } from './notifications/routes.js'
 import { serviceMaintenance } from './maintenance/routes.js'
-import { serviceResources } from './resources/routes.js'
 
 export const service = {
   plugin: {
@@ -16,7 +15,6 @@ export const service = {
         serviceAutomations,
         serviceNotifications,
         serviceProxy,
-        // serviceResources,
         serviceSecrets,
         serviceMaintenance,
         serviceTerminal
